Check format of dates in CSV import file
Add code to check that dates in the input data in the CSV are in the
expected format. All possible date fields are checked, since the
expected format (dd/mm/yy) can sometimes match US format (mm/dd/yyyy):
If one date is wrong, it may be that all dates are in the wrong (esp.
US) format, so important to fail early before any data actually
imported to avoid adding data with broken dates.