Limit dataset size
Imports of JG data fails if too many records are added at once: The
results of the import from assist_just_giving_import() are a long
string which is saved to the database, so this can't be too long.
To prevent this, imports are limited to files of <= 100 records.

This is a quick-and-dirty fix, better would be to cut this results
string into smaller chunks and save each of them.